python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "autogen-core>=0.4.3",
  "autogen-ext[anthropic,azure,mcp,ollama,openai]>=0.4.3",
  "autogen-agentchat>=0.4.3",
  "autogenstudio>=0.4.3",
  "openai>=1.72.0",
  "tiktoken==0.8.0",
  "python-dotenv>=1.0.1",
  "pyyaml>=6.0.2",
  "mcp>=1.2.0",
  "sqlite-vec>=0.1.0",
  "typer>=0.15.2",
  "numpy>=2.2.3",
  "opentelemetry-api>=1.30.0",
  "opentelemetry-sdk>=1.30.0",
  "opentelemetry-exporter-otlp-proto-grpc>=1.30.0",
  "opentelemetry-instrumentation-openai>= 0.38.0",
  "opentelemetry-instrumentation-httpx >= 0.51.0",
  "anthropic>=0.49.0",
  "pinecone>=6.0.2",
  "boto3>=1.34.41",
  "litellm>=1.30.7",
  "fastapi>=0.110.0",
  "pydantic>=2.0.0",
  "httpx>=0.25.0",
]

[project.optional-dependencies]
jupyter-executor = [
    "ipykernel>=6.29.5",
    "nbclient>=0.10.2",
]
test = [
  "pytest>=8.3.5",
  "pytest-asyncio>=0.25.3",
]
lint = [
  "ruff>=0.9.2",
]

[project.scripts]
kagent-engine = "kagent.cli:run"
tool_gen = "kagent.tools.utils.tool_gen:main"

[tool.uv.sources]
autogenstudio = { git = "https://github.com/kagent-dev/autogen.git", subdirectory = "python/packages/autogen-studio", rev = "ec221e39067a8baeb117e7bf5e3d04343d0cb7ad" }
autogen-agentchat = { git = "https://github.com/kagent-dev/autogen.git", subdirectory = "python/packages/autogen-agentchat", rev = "ec221e39067a8baeb117e7bf5e3d04343d0cb7ad" }
autogen-ext = { git = "https://github.com/kagent-dev/autogen.git", subdirectory = "python/packages/autogen-ext", rev = "ec221e39067a8baeb117e7bf5e3d04343d0cb7ad" }
autogen-core = { git = "https://github.com/kagent-dev/autogen.git", subdirectory = "python/packages/autogen-core", rev = "ec221e39067a8baeb117e7bf5e3d04343d0cb7ad" }
kagent = { workspace = true }

[tool.ruff]
line-length = 120
fix = true
exclude = [
  "notebooks/",
]

target-version = "py312"

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
select = ["E", "F", "W", "B", "Q", "I", "ASYNC", "T20"]
ignore = ["F401", "E501"]

[tool.hatch.build.targets.wheel]
packages = ["src/kagent"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
python_files = "test_*.py"
python_functions = "test_*"
log_cli = true
log_cli_level = "INFO"
